+++ resources.go
@@ -145,6 +145,12 @@ func Provider() tfbridge.ProviderInfo {
 			Requires: map[string]string{
 				"pulumi": ">=1.0.0,<2.0.0",
 			},
+            Overlay: &tfbridge.OverlayInfo{
+                    DestFiles: []string{
+                            "pulumi_docker/docker.py",
+                            "pulumi_docker/image.py",
+                    },
+            },
 		},
 		CSharp: &tfbridge.CSharpInfo{
 			PackageReferences: map[string]string{